Make serialization of Snowflake consistent
Snowflakes are decoded as Longs but encoded as Strings. Mixing these types is allowed in JSON (and is done by Discord quite frequently) but does lead to other encoding types failing.

This commit changes snowflakes to be encoded and decoded as long values exclusively, and should fix the issue.

+++ gateway/src/test/kotlin/json/CommandTest.kt
@@ -70,7 +70,7 @@ class CommandTest {
         val json = json.encodeToString(JsonObject.serializer(), buildJsonObject {
             put("op", OpCode.RequestGuildMembers.code)
             put("d", buildJsonObject {
-                put("guild_id", guildId)
+                put("guild_id", guildId.toLong())
                 put("query", query)
                 put("limit", limit)
             })
@@ -95,8 +95,8 @@ class CommandTest {
         val json = json.encodeToString(JsonObject.serializer(), buildJsonObject {
             put("op", OpCode.VoiceStateUpdate.code)
             put("d", buildJsonObject {
-                put("guild_id", guildId)
-                put("channel_id", channelId)
+                put("guild_id", guildId.toLong())
+                put("channel_id", channelId.toLong())
                 put("self_mute", selfMute)
                 put("self_deaf", selfDeaf)
             })
